+++ contracts/src/v0.8/interfaces/OperatorInterface.sol
@@ -15,7 +15,7 @@ interface OperatorInterface is OracleInterface, ChainlinkRequestInterface {
     bytes calldata data
   ) external;
 
-  function fulfillOracleRequest2(
+  function fulfillOperatorRequest(
     bytes32 requestId,
     uint256 payment,
     address callbackAddress,